move observation to modules:observation
