+++ modules/observation/src/main/java/org/example/event/sourcing/order/poc/modules/observation/annotation/LogInfo.java
@@ -1,4 +1,4 @@
-package org.example.event.sourcing.order.poc.observation.annotation;
+package org.example.event.sourcing.order.poc.modules.observation.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
