+++ modules/observation/src/main/java/org/example/event/sourcing/order/poc/modules/observation/aop/DefaultObserveAroundMethodHandler.java
@@ -1,4 +1,4 @@
-package org.example.event.sourcing.order.poc.observation.aop;
+package org.example.event.sourcing.order.poc.modules.observation.aop;
 
 public class DefaultObserveAroundMethodHandler extends AbstractObserveAroundMethodHandler {
 }
