+++ modules/observation/src/main/java/org/example/event/sourcing/order/poc/modules/observation/config/LogAspectConfig.java
@@ -1,7 +1,7 @@
-package org.example.event.sourcing.order.poc.observation.config;
+package org.example.event.sourcing.order.poc.modules.observation.config;
 
-import org.example.event.sourcing.order.poc.observation.aop.AbstractLogAspect;
-import org.example.event.sourcing.order.poc.observation.aop.DefaultLogAspect;
+import org.example.event.sourcing.order.poc.modules.observation.aop.AbstractLogAspect;
+import org.example.event.sourcing.order.poc.modules.observation.aop.DefaultLogAspect;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
