+++ modules/observation/src/main/java/org/example/event/sourcing/order/poc/modules/observation/config/ObserveConfiguration.java
@@ -1,9 +1,9 @@
-package org.example.event.sourcing.order.poc.observation.config;
+package org.example.event.sourcing.order.poc.modules.observation.config;
 
 import io.micrometer.observation.ObservationRegistry;
 import io.micrometer.observation.aop.ObservedAspect;
-import org.example.event.sourcing.order.poc.observation.aop.AbstractObserveAroundMethodHandler;
-import org.example.event.sourcing.order.poc.observation.aop.DefaultObserveAroundMethodHandler;
+import org.example.event.sourcing.order.poc.modules.observation.aop.AbstractObserveAroundMethodHandler;
+import org.example.event.sourcing.order.poc.modules.observation.aop.DefaultObserveAroundMethodHandler;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
