+++ modules/observation/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,2 @@
+config.org.example.event.sourcing.order.poc.modules.observation.ObserveConfiguration
+config.org.example.event.sourcing.order.poc.modules.observation.LogAspectConfig
