--- observation/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,0 @@
-org.example.event.sourcing.order.poc.observation.config.ObserveConfiguration
-org.example.event.sourcing.order.poc.observation.config.LogAspectConfig
