+++ order/event-handler/build.gradle
@@ -4,7 +4,7 @@ dependencies {
     implementation project(":common")
     implementation project(":modules:client")
     implementation project(":event")
-    implementation project(":observation")
+    implementation project(":modules:observation")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter'
