+++ settings.gradle
@@ -2,12 +2,13 @@ rootProject.name = 'event-sourcing-order-poc'
 
 include 'common'
 include 'event'
-include 'observation'
 include 'modules'
 include 'modules:idempotency'
 findProject(':modules:idempotency')?.name = 'idempotency'
 include 'modules:client'
 findProject(':modules:client')?.name = 'client'
+include 'modules:observation'
+findProject(':modules:observation')?.name = 'observation'
 
 include 'order'
 include 'order:command-side'
