+++ shipment/command-side/src/main/java/org/example/event/sourcing/order/poc/command/shipment/service/ShipmentService.java
@@ -5,7 +5,7 @@ import org.example.event.sourcing.order.poc.command.shipment.producer.ShipmentEv
 import org.example.event.sourcing.order.poc.common.model.Shipment;
 import org.example.event.sourcing.order.poc.event.model.ShipmentEvent;
 import org.example.event.sourcing.order.poc.event.model.ShipmentEventName;
-import org.example.event.sourcing.order.poc.observation.annotation.LogInfo;
+import org.example.event.sourcing.order.poc.modules.observation.annotation.LogInfo;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
