+++ shipment/query-side/build.gradle
@@ -2,7 +2,7 @@ jib.to.image = "noahhsu/espoc-shipment-${project.name}"
 
 dependencies {
     implementation project(":common")
-    implementation project(":observation")
+    implementation project(":modules:observation")
     implementation project(":event")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
